+++ gosnmp.go
@@ -428,12 +428,14 @@ func (x *GoSNMP) SnmpEncodePacket(pdutype PDUType, pdus []SnmpPDU, nonRepeaters 
 
 	pkt := x.mkSnmpPacket(pdutype, pdus, nonRepeaters, maxRepetitions)
 
-	// Request ID is an atomic counter (started at a random value)
-	reqID := atomic.AddUint32(&(x.requestID), 1) // TODO: fix overflows
+	// Request ID is an atomic counter that wraps to 0 at max int32.
+	reqID := (atomic.AddUint32(&(x.requestID), 1) & 0x7FFFFFFF)
+
 	pkt.RequestID = reqID
 
 	if x.Version == Version3 {
-		msgID := atomic.AddUint32(&(x.msgID), 1) // TODO: fix overflows
+		msgID := (atomic.AddUint32(&(x.msgID), 1) & 0x7FFFFFFF)
+
 		pkt.MsgID = msgID
 
 		err = x.initPacket(pkt)