+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * [ENHANCEMENT]
 * [BUGFIX]
 
+* [BUGFIX] Add bounds checking for reqID and msgID #273
+
 ## v1.29.0
 
 NOTE: This release returns the OctetString []byte behavior for v1.26.0 and earlier.